Task1/requirements.txt
streamlit
tensorflow>=2.10
opencv-python
Pillow
numpy
langdetect
transformers
torch
sentencepiece
tqdm
